Changed icons to use Font-Awesome.
Using Font-Awesome is easier as well as requires less codes than the previous way of implementing icons which we had to create a lot of files for each individual folder.
In the current implementation, we're storing all icons in a single file 'Icon.js' and importing as necessary in components.